refactor: relocate legacy loader subsystem to services, repoint workers
Phase 12F. Two parts.

1. Bootstrap re-points (safe — old modules are re-export shims):
   - components.indexer.loaders.serializer.DocSerializer
       -> services.workers.parsers.doc_serializer
   - components.indexer.loaders.pdf_loaders.docling2.DoclingPool
       -> services.workers.parsers.docling_workers
   - components.indexer.loaders.pdf_loaders.marker.MarkerPool
       -> services.workers.parsers.marker_workers

2. Loader-registry relocation. The plan assumed the loaders were already
   migrated to core/indexing/parsers/ and the old ones were dead. They are
   not: core/indexing/parsers/registry.py is an empty Registry, and the new
   doc_serializer / doc_serializer_bridge still drive the legacy loaders via
   get_loader_classes (which walks the loaders package and discovers ~20
   BaseLoader adapters). So the legacy loader subsystem is the live runtime
   path. To make components/ deletable (12H) it is moved wholesale:

   components/indexer/loaders/  ->  services/workers/parsers/legacy_loaders/

   - get_loader_classes root_pkg updated to the new package path.
   - The two get_loader_classes consumers (doc_serializer, doc_serializer_bridge)
     and internal absolute imports / tests repointed.
   - The orphaned serializer.py re-export shim (no remaining consumers after the
     bootstrap re-point) is deleted rather than carried into services/.
   - Internal `from components.utils` / `from components.prompts` imports inside
     the loaders are left as-is; they remain valid and are 12C/12E/12H's concern.

   Runtime discovery verified: get_loader_classes resolves all 19 extension
   mappings from the new location.

Also fix a latent test-isolation bug this relocation surfaced:
services/inference/parsers/test_openai_audio.py installed a fake non-package
`pydub` into sys.modules at import time and never restored it. It was masked
only by collection order (the real-pydub loader test used to sort before it
under components/). It now tries the real pydub first and only falls back to
the stub when the import genuinely fails — its stated Python-3.13 purpose.

No remaining `components.indexer.loaders` references anywhere.
