+++ openrag/services/workers/parsers/legacy_loaders/CustomDocLoader.py
