+++ openrag/services/workers/parsers/legacy_loaders/image.py
