+++ openrag/services/workers/parsers/legacy_loaders/pdf_loaders/docling2.py
