+++ openrag/services/workers/parsers/legacy_loaders/pdf_loaders/dotsocr.py
