+++ openrag/services/workers/parsers/legacy_loaders/test_eml_recursion.py
@@ -45,7 +45,7 @@ def _make_leaf_eml() -> bytes:
 
 @pytest.mark.asyncio
 async def test_eml_recursion_caps_at_max_depth(tmp_path):
-    from components.indexer.loaders.eml_loader import EmlLoader
+    from services.workers.parsers.legacy_loaders.eml_loader import EmlLoader
 
     # A single outer .eml whose attachment is another .eml. We seed the
     # call at depth = cap - 1, so processing the outer's attachment (which
@@ -69,7 +69,7 @@ async def test_eml_recursion_caps_at_max_depth(tmp_path):
 @pytest.mark.asyncio
 async def test_eml_below_cap_does_not_skip(tmp_path):
     """At depth 0 the guard does not fire — attachments are still attempted."""
-    from components.indexer.loaders.eml_loader import EmlLoader
+    from services.workers.parsers.legacy_loaders.eml_loader import EmlLoader
 
     eml_path = tmp_path / "nested.eml"
     eml_path.write_bytes(_make_eml_with_attached_eml(_make_leaf_eml()))
