+++ openrag/services/workers/parsers/legacy_loaders/txt_loader.py
@@ -14,13 +14,13 @@ migrate.
 import asyncio
 from pathlib import Path
 
-from components.indexer.loaders.base import BaseLoader
 from core.indexing.parsers.markdown_parser import MarkdownParser
 from core.indexing.parsers.text_parser import TextParser
 from core.models.document import Document as CoreDocument
 from core.models.document import DocumentType
 from core.utils.logging import get_logger
 from langchain_core.documents.base import Document
+from services.workers.parsers.legacy_loaders.base import BaseLoader
 
 logger = get_logger()
